backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pdfplumber
python-docx
openai
python-dotenv